iwinfo: add patch to properly handle per_sta_vif case
When per_sta_vif is enabled in hostapd BSS config, we need to iterate
through every interface in /sys/class/net which matches
`${interface_requested}.${number}`. Otherwise data requested from iwinfo
will be incomplete/nonexistant. This is already handled for WDS so there is
precedence for a change like this.

Also refactor logic on whether to issue a NL80211_CMD_GET_STATION request
on the interface into a seperate function.

Signed-off-by: rany <ranygh@riseup.net>