mac80211: mtk: fix incorrect VIF assignment for IEEE 802.11 fragments (3ae8393)
When in WDS mode (vif is AP_VLAN), only the first fragment would be dequeued from
txq and go through some tx processing which includes the replacement of vif in CB(SKB)
from the AP_VLAN one to the AP one. However, the rest of the fragments would be dequeued
from txqi->frags and bypass this process which in turn results in different VIF assignment
for softmac to fill in TXD. Therefore, we fix this by assigning the correct AP VIF to
SKB_CB of the rest fragments.

Signed-off-by: Rany Hany <rany_hany@riseup.net>